Adjusted auditor unit test for inconsistencies.
Auditor is only intended to return an inconsistencies report if the
portfolios are different. What those inconsistencies look like depends
solely on the behavior of Portfolio, which contains the method to
actually return the inconsistencies. Therefore the nature of the
inconsistencies returned shall be tested in the Portfolio unit test.