Adjusted auditor unit test to handle database access failure.
Auditor is not guaranteed to successfully access the external database.
However, it has been previously defined that Auditor is responsible for
handling such a situation, via providing a report that indicates the
problem. The intent is for the error to be non-fatal, and provide an
opportunity for the operator to rectify the situation. The needed unit
test along with try/catch blocks were added.