Further design changes to ProgramManager made.
Changes made mainly include changing runUserInputCommand to use
the argParse wrapper closer inline to the libraries intended use - that is,
preparation of the operation should be done by the caller, not the wrapper.

It will soon be time to move onto implementing supported operations.
Currently the only supported operation is that which stops Erasmus. Some
changes were also made to Main, namely a null check as NextLine() from
scanner will not return null (even though we want this to happen).