Cleanup to BasicMap and ArgumentParserWrapper performed.
The intent of these wrappers - to avoid polluting the Erasmus code with
3rd party API calls - was not entirely clarified before. Namely, should
the functionality of the API be matched exactly, or alterations
allowed for the purpose of conciseness?

As of now, small alterations at my discretion will be permitted. If
these alterations result in a loss of needed generality, I can always
refactor later. I will otherwise stay faithful to the API as much as
possible.