Learner tests to determine command nesting ability created.
The commands of Erasmus will eventually be numerous and varied. Thus
it is necessary to support deeply nested sub commands which can accept
appropriate arguments. A learner test to determine whether or not
subparsers can have further subparsers was created for this purpose.
It turns out they can. Further work needs to be done on the wrapper
classes to make this infinite nesting ability more streamlined.