Refactored ArgParseWrapper to combine parsers and subparsers.
This helps remove some of the complexity of the underlying library,
but unfortunately I now know this removes some of the type safety that
argParse4j inherently provides in its fluent syntax. This is perhaps
all the more reason to wrap more faithfully, and THEN make my adapters.
I may go back to wrapping more thinly if null pointers are too common.