Some class naming cleanup done, and learner tests created.
A learner test for the arg parse library to determine how top level
and sub commands work together was created. It was determined that
top level arguments will be added to the map alongside whatever
the sub parser adds itself. Additionally, an important discovery was
made: the library cannot easily differentiate sub commands from a 
list of positional arguments if the number of arguments can be arbitrary.
This is obviously to prevent command ambiguity, but limits flexibility.
It will be necessary to keep this in mind when designing the commands.

On a smaller note, it turns out I implemented equals() incorrectly for
my util wrapper classes, something I found out when creating the
learner tests outlined. I have since fixed that.