ArgParseWrapper now automatically creates the subparsers object.
Null pointer exceptions due to tests or scenarios calling the wrapper
without the subparsers explicitly being created (as is expected in the
original library) were proving to be very problematic. Auto creation
helps tremendously. Also, the old and incorrect implementation of equals()
on BasicArrayList was considering an ArrayList equal to BasicArrayList,
which in turn caused some bugs in tests where a wrapper was not
correctly used.