Added some arguments and corresponding unit tests for Program Manager.
Earlier, unit tests were set up to allow for new commands to be added
and tested incrementally. This marks the completion of the first command
interface, albeit without the process itself supported yet. The test
was successful as the proper map argument was passed by the program.
This map will later be used to properly set up the process and execute
as intended.