Learner tests for JFreeChart library created.
Not all program manager commands are complete yet, but it is vital to
get an end-to-end function working before creating a wide breadth of
commands. Hence progressing to the learner tests for JFreeChart is the
next step. Also, I found out that downloading the javadocs and source
code for these open source libraries and integrating it with NetBeans
is very simple. I have done so for the purpose of speeding up the
project.