Refactoring of all internally used utilwrappers.
The wrappers of the collections classes and maps were refactored to
be more faithful to the contracts of the java util wrappers whilst
still allowing for simple refactoring should collections ever change.
This was a considerable learning experience, and I had to carefully
define many boundaries, but this will almost certainly help considerably
later on.