Continued work on the plotter class.
It was found necessary to specify the desired header of the csv file.
Thus some modifications were made to ProgramManager as well. IndexOf
was also added to the list wrapper, and unlike the underlying class,
throws an exception instead of returning the nebulous error code index of -1
if the item is not present.