Refactoring of file reader to permit for generalization.
Note: Plotter tests are not passing. Refactoring was done to the
basic file reader to generalize the functionality to allow for other
data sources. The plotter is also partially underway to be adapted to
accept different formats, as database input information will be
formatted differently than the data read by from the NYSE file data.