ParsedArgs now uses an Enum for keys, instead of strings.
Usage of strings as named constants for keys in the maps provided by
the command line parser have been an ongoing problem in ensuring the
desired fields are being referenced. The command line parser library
wrapper now returns the map with Enums as keys, and all code has been
updated with this assumption.